HIVD_classifier/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
torch>=2.1.0
torchvision>=0.16.0
open-clip-torch>=2.24.0
ultralytics>=8.1.0
albumentations>=1.3.1
pillow>=10.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
redis>=5.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
websockets>=12.0
pyyaml>=6.0
tqdm>=4.66.0
aiofiles>=23.2.0
